requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
pydantic==2.12.5
openai==2.20.0
python-dotenv==1.2.1
loguru==0.7.3
pytest==9.0.2
